fix(sam): numpy compatibility et serialisation des masks
- Downgrader numpy à <2.0 pour compatibilité avec segment-anything+torch
- Installer numpy AVANT torch dans le Dockerfile pour éviter override
- Convertir les masks (bool) et IOU predictions (float) en int/float avant .tolist()

Problèmes résolus:
- 'Could not infer dtype of numpy.uint8' → NumPy 2.x incompatible avec SAM
- Validation errors on masks (bool vs int) → astype(int) avant serialisation
- Dockerfile COPY cache était doublon (docker/sam_server.py vs jobs/sam/sam_server.py)

+++ jobs/sam/sam_server.py
@@ -193,8 +193,8 @@ async def _segment_impl(request: SegmentationRequest) -> SegmentationResponse:
 
         return SegmentationResponse(
             status="success",
-            masks=masks.tolist() if masks is not None else None,
-            iou_predictions=iou_predictions.tolist() if iou_predictions is not None else None,
+            masks=(masks.astype(int).tolist() if masks is not None else None),
+            iou_predictions=(iou_predictions.astype(float).tolist() if iou_predictions is not None else None),
         )
 
     except Exception as e:
